+++ src/assets/styles/allcategory.css
